rename file with test plans config
